English直後のToonTextureで読み込みに失敗する

+++ Format/PMD/IK.cs
@@ -7,7 +7,7 @@ namespace MMD
     {
         namespace PMD
         {
-            public class IKList : ChunkList<IK, uint> 
+            public class IKList : ChunkList<IK, ushort> 
             {
                 public List<IK> IKs { get { return elements; } }
             }
